SaveChangeAsync Interceptor

+++ src/Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -8,7 +8,7 @@ namespace ETicaretAPI.Domain.Entities
 
         public int Stock { get; set; }
 
-        public long Price { get; set; }
+        public float Price { get; set; }
 
         public ICollection<Order> Orders { get; set; }
     }
